refactored usecase classes
