+++ README.md
@@ -1,7 +1,7 @@
 # CSC207 Project: Lost-the Game
 
 As of Nov. 29, we have finished most of our features for the game. We just need to finish the UI and one small feature 
-between the CombatInteractor class and the usecases.StorylineInteractor class. After these implementations, our program will 
+between the CombatInteractor class and the usecases.story.StorylineInteractor class. After these implementations, our program will 
 be able to run. 
 
 ## Progress Report
