+++ implementation_plan.md
@@ -4,6 +4,6 @@
 - **Chris** : `Event` & `CombatEvent` & `EnemyData` & `QuestionsData`, `CombatInteractor`, `EventManager` *(end of Oct 29)*
 - **Natasha** : `entities.player.PlayerData`, `entities.player.PlayerInteractor` *(end of Oct 29)*
 - **Alaina** : `LoadInteractor`, `<LoadInterface>`, `SaveInteractor`, `<SaveInterface>`
-- **Josh** : `usecases.StorylineInteractor`, `<StorylineInterface>`
+- **Josh** : `usecases.story.StorylineInteractor`, `<StorylineInterface>`
 - **Ludan** : `ViewController`, `View`
 - **Gwen** : `SoundInteractor`, `LoginInteractor`
