+++ src/main/java/entities/player/PlayerFactory.java
@@ -1,4 +1,4 @@
-package usecases;
+package entities.player;
 
 import entities.items.ItemData;
 import entities.player.PlayerData;
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 
 public class PlayerFactory {
-    static PlayerData generatePlayer(String username, int eventID, int ap, HashMap<String, ArrayList<ItemData>> inventory) {
+    public static PlayerData generatePlayer(String username, int eventID, int ap, HashMap<String, ArrayList<ItemData>> inventory) {
         return new PlayerData(username, eventID, ap, inventory);
     }
 }
