+++ src/main/java/usecases/gamesave/Loader.java
@@ -1,4 +1,4 @@
-package usecases;
+package usecases.gamesave;
 
 import entities.player.PlayerData;
