+++ src/main/java/usecases/login/LoginInteractor.java
@@ -1,4 +1,4 @@
-package usecases;
+package usecases.login;
 
 import presenter.LoginInterface;
